requirements.txt
alembic==1.18.4
amqp==5.3.1
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.3
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.3
ordered-set==4.1.0
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.2
WTForms==3.2.1
apscheduler==3.10.4